document `--rm` semantics
The `--rm` flag will only cause a container to be removed when it has
been created and started successfully.  Otherwise, it will not be
removed to allow the container to be inspected and to analyze the root
cause of the failure.  Document those semantics more clearly in the
manpages to avoid confusion for users.

Fixes: #1359
Signed-off-by: Valentin Rothberg <vrothberg@suse.com>

Closes: #1362
Approved by: rhatdan