add install and uninstall commands
add install and uninstall commands to mimic the atomic cli commands so users
can use podman in place of the atomic cli.

the install command leverages an INSTALL label that has an embeddded command to
run to prepare the OS or container for the system.

the uninstall command leverges an UNINSTALL label that has an embedded command
that cleans the system including deleting the image (and its containers).

Signed-off-by: baude <bbaude@redhat.com>