error when exec'ing a command with no tty
Due to the way container tty access is currently wired in podman, we
must have a tty available where the user executes podman.  Failure to
do so results in a hard podman lock where no container related commands
can be run.

When we can leverage conmon for console handling, this requirement can
be removed.

Signed-off-by: baude <bbaude@redhat.com>