Run stop in parellel
Currently when stopping a container we can take 10 seconds per container to stop
but we do this synchronously.

Running the containers stops in parrallel can speed up the stop by N times,
where N is the number of containers we are stopping.

Tested starting and stopping 3 sleep containers, time went from 30 seconds
to 10 seconds.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>